Init cache manager with customer config when both customer and classpath config exist (#1142)
When try to enable distributed cache, I found current way may cause
ISPN000343 problem. Because distributed cache needs transport support in
global config, so if the cache manager does not enable it and then add
this type of cache by cacheManager.defineConfiguration, it will report
this error. So this commit is a workaround fix for the problem.
  With this fix, when we use customer+classpath config way, we should
enable <transport/> in customer config first, and then add any
distributed caches in both these configurations.