upgrade weft to fix threadlocal and add diag client
The diags client currently just summarizes thread counts using a rough approximation of thread group names.

The weft upgrade fixes a problem with the ThreadContext clear method, where the memory wasn't being released.

Also adding a minor refactor to MavenMetadataGenerator to pass the incremental state objects instead of their parts,
which simplifies the method sig and the code that calls it.