Add JNDI data sources
This will help us reuse the same DataSource in all Infinispan caches that get persisted to the DB.

Basically, this allows us to configure a datasource in the new conf.d/connection-pools.conf file,
which will be bound to a JNDI context that is only available within Indy, then lookup those
datasources elsewhere in the application.

Note that, if we use the old style of BootupAction implementation that's part of a class which
uses an Infinispan cache, it may trigger Infinispan to initialize before the datasource is actually
bound!