switch to propulsor configuration, remove unused add-ons
I switched to ConfigUtils.loadStandardConfigProperties(), which includes environment variables
under the 'env.' prefix. This is new in Commonjava/propulsor#37.

Switched over to using propulsor configuration so we can continue to consolidate onto that as
a platform for our services (for as long as that makes sense). This is the bulk of the changes
here, but I'll mark the things that aren't related when I create the PR.

Changed logging of 404's in ResponseUtils to use logger.debug, but other errors are still
logged as error.

Removed autoprox, setback, and relate, since these are all associated with use cases that
no longer exist for Indy and they were just cluttering up the codebase / complicating the
logic.