switch to propulsor configuration, remove unused add-ons (#1230)
* switch to propulsor configuration, remove unused add-ons

I switched to ConfigUtils.loadStandardConfigProperties(), which includes environment variables
under the 'env.' prefix. This is new in Commonjava/propulsor#37.

Switched over to using propulsor configuration so we can continue to consolidate onto that as
a platform for our services (for as long as that makes sense). This is the bulk of the changes
here, but I'll mark the things that aren't related when I create the PR.

Changed logging of 404's in ResponseUtils to use logger.debug, but other errors are still
logged as error.

Removed autoprox, setback, and relate, since these are all associated with use cases that
no longer exist for Indy and they were just cluttering up the codebase / complicating the
logic.

* implement indy.node.id config param and inject into system props

Also adjust ResponseUtils logging to log error for code > 499, otherwise debug.

* update AuditQuery publisher to use indy.node.id from config object

* Rename SystemPropertyProvider.getSystemProperties() to make more intuitive, and fix compile error

* Code change for schedule manager cluster support

  This commit contains following:

  * New cache to serve as cluster lock of schedule expiration following
actions in cluster
  * Schedule cache listener control logic in cluster support
  * Remove some useless conf for quartz legacy

Conflicts:
	core/src/main/java/org/commonjava/indy/core/expire/ScheduleManager.java

* Use ScheduleKey instead of UUID as schedule-lock-cache key

* Add cluster configuration of ispn

Conflicts:
	subsys/infinispan/src/main/resources/infinispan-cluster.xml

* Fix infinispan upgrade of 9.x

Conflicts:
	subsys/infinispan/src/main/resources/infinispan-cluster.xml

* Use indy-global nodeId and change producer for this node to LocalIndyNodeProvider

* fix a couple of stupid bugs