Refine metric handling and create REST endpoint MDC value
I combined the class- and method-level @Path annotation values to create a combined REST endpoint
MDC value that we can log, which makes the division between levels less arbitrary when seen from logs.

I also added MDC / logging support for metrics reported via wrapWithStandardMetrics, then
refined the way these meters / timers are managed to try to consolidate duplicated code a bit.

Finally, I reduced the number of possible parameters in @Measure to the things we actually use.
I'd like to remove these parameters altogether, but can't quite rule out the notion that we might
need a secondary / supplemental metric name that spans multiple sites in the code. For now, I
force the calculation and injection of the default metric names in addition to any that might
be specified, to make the metrics schema a bit more predictable when in the presence of
custom names.