[Draft] Implement 4 golden signals directly in Indy metrics
This approach will use a servlet filter to wrap REST traffic and analyze it using previously
established metadata (currently added to the MDC, but eventually copied to ThreadContext).
Then this analysis will be used to classify the Indy function that the request accessed, and
log metrics by function. Metrics will be limited to latency, errors, and throughput. Saturation
will be measured separately for now...until we understand it better.

I still need to add the actual classification logic, and copy the classification information
from the MDC into the ThreadContext. Using MDC for anything except logging is kind of an ugly hack.