Update NPM support to do deeper metadata path mapping
When accessing package metadata for NPM, we need separation between storage location and logical / upstream
path. This allows us to avoid colliding between the metadata's requested path and the directory structure
used to access tarballs. We had been doing this mapping in certain cases, in the NPMContentAccessHandler.

However, this resulted in weird behavior when the metadata files expired, and the directory structure
was left in place, but empty.

This reimplementation provides a new interface for calculating storage paths, used by Indy's custom
path generator (which in turn is used in the CacheProvider from Galley). Then, we implement this
interface in pkg-npm/common, in order to handle the NPM specific case.

Finally, an upstream-server expectation was commented out of the NPMMetadataTimeoutTest in order
to simulate real-world usage.