Switch to StreamingOutput impl for DTO rendering
This uses a form of StreamingOutput as a response entity in order to avoid measuring rendering / transfer times for DTOs in with latency, since these are not logical functions.

In the process, I've added logic to wrap DTO rendering / streaming with metrics for this process as well, in case we need it for diagnosis. This required switching ResponseUtils to a helper class, ResponseHelper, which is CDI injected into REST resources.