Make NPM JSON tarball replacement parse path tail to handle subpaths
We have some cases where internal registry caches have more complex URLs, with subpaths. In these cases, simply parsing the URL and pulling the path, then appending to the Indy version of the URL will not work correctly (it will result in longer paths than it should).

To fix this, I divide the URL path and pull parts from the end to determine what the NPM content path is. Then, I reconstruct that path from the parts and append it to the Indy URL in the metadata.

I'm not completely sure I've accounted for all use cases here.