Fix ftest failure
  When a pom file changed in a hosted repo, it will trigger a
maven-metadata removing in the GA level and cascading removing for
affected groups. For fix problem of this, we use DownloadManager.delete
to solve readonly hosted repo deletion not allowed problem, but it does
not support clear group metadata. So this fix will allow remove the
mergable metadata through that method.