Add cache listener on koji maven metadata to clean things up
Currently, we rely on cache expirations to trigger cleanup of metadata files
(at least in the absence of content changes on Indy itself). This is fine, except in
the case where Maven metadata is cached for a group, which was constructed using
metadata derived from Koji builds. In this case, the Koji metadata object is cached
with some expiration, but then so is the aggregated group metadata.

What happens is that the Koji metadata object expires, which should trigger Indy to
re-query Koji for updated metadata. However, because that metadata info has been
incorporated into the aggregated form of the metadata at the group level, the caching
at the group level keeps getting reset with each access. This prevents expiration of
the group-level metadata, which means we never actually go back to the Koji metadata
provider and do that re-query.

This change introduces a CacheListener on the Koji metadata cache, such that when
an entry expires there, it finds the groups where Koji metadata is enabled, and
expires the aggregated metadata for those groups and any affected groups. This will
not only clear the cached group metadata for the file, but also the aggregated file
itself.

This implementation reuses the logic used to clear metadata paths when group
membership changes, in order to provide a central point for refactoring that logic.