Improved metric names for transfer decorators, more metrics for promotion
I switched to using IndyMetricsManager for the metrics used in transfer decorators. For decorators in Indy itself,
I just inject the metrics manager, but for decorators in the galley codebase I'm using a Function passed in,
which allows DefaultGalleyStorageProvider to use the metrics manager to retrieved timers with proper names
for code that's actually in galley.

I also annotated more methods in DefaultContentManager and PromotionValidationTools with @Measure,
to give us some more fine-grained APM measurements around the promotion process.

Finally, I added metric timers to the raw transfer decorator, and made it wrap InputStream's too.
This will allow us to get APM measurements for the raw process of reading/writing to the filesystem.
It's important to note that these timers will be influenced by things like network transfer rates
(when writing uploaded files) and CPU / thread congestion (when reading or writing files).