Add an in memory facade before ispn store manager
   Now in ISPNDataStoreMgr, we are storing the store data as json
string, and when querying we need to deser the json to object each time.
As the query for store is a high frequent ops, this may cause some time
wasting, so we need to think about to use the store object directly in.
   To avoid object externalizing incompatibility problem, I think we
should not change the way of ispn marshalling/unmarshalling from
persistence. So I added a in memory facade map between caller and
underlying ispn cache. As this store data cache is not a big one, so I
think it will not bring memory problem here.