Switch to using Externalizable/ArtifactStore for InfinispanStoreDataManager (#1455)
* Switch to using Externalizable/ArtifactStore for InfinispanStoreDataManager

Previously, storing ArtifactStore definitions as JSON requires us to serialize/deserialize
(render and parse) constantly, every time we access a repository definition. We can access
dozens or even thousands of these in a request, especially in an upload request or a
promotion request.

Looking at our current production deployment, we see that upload latency is up around 7 seconds
at times, with about 6.98 seconds of that being consumed in the process of working with the
store-definition cache.

This change implements Externalizable for all ArtifactStore derivatives, StoreKey, and related
classes in the core model of Indy. It adds some testing to ensure these can be read / written
appropriately using readExternal / writeExternal methods, and verifies that round-trip data is
preserved.

Then, it switches the cache used by InfinispanStoreDataManager to use Cache<StoreKey, ArtifactStore>
instead of Cache<StoreKey, String>, and relies on Infinispan itself to use the Externalizable
implementations to serialize data when it is persisted to the backing database.

Along with this comes a need to migrate existing JSON data to the new binary format. To handle
this, I've modified the InfinispanStoreDataMigrationAction class to look for a cache named
'store-data-v1' and if it finds it, to migrate the old-style store definitions from JSON to
the new cache, managed by InfinispanStoreDataManager. It will NOT use the really old filesystem
storage location if it finds this legacy cache.

* make no-arg constructors public and correct usage of Object[In/Out]putStream

* Add mixed-lot serialization for ArtifactStore, and read all by ArtifactStore base class

This will test that polymorphism works in the object stream.