Using store-data-v1 for new object-based store-data cache (#1456)
I found a problem for migration process. When we use db-based
persistence for store cache, we can not control the table name for which
store use by applying a simple "table-name" like attr in ispn xml
because it does not support this. The table name is applied by its cache
name directly. So we need to use a new cache name for the new obj-based
store, ane use the legacy "store-data" cache for migration purpose.