Rewrite store-query methods that use stream() and intercept/wrap store-query for metrics
We cannot use the @Measure annotation in ArtifactStoreQuery, since that class is instantiated a variable number of
calls for each user request, and instances hold state supporting the fluent API. The StoreDataManager decorator
wraps ArtifactStoreQuery results from the query() method in a query class that measures query calls, delegating the
actual query to the instance that comes from the underlying StoreDataManager delegate.

StoreDataManager stream / getAll methods rely on having in-memory access to ALL of the ArtifactStore instances,
and also on having a small enough instance count to make iterating through all of them feasible. In our production
environment, both of these assumptions are violated. We need a better way to access this information, but to start
I've refactored DefaultArtifactStoreQuery methods that use streams and getAll in the StoreDataManager so they won't
do that.

In the case of retrieving the stores that are in the membership tree of a particular group, I've rewritten that to
avoid recursion.

In addition, I've added a keyStream() method to ArtifactStoreQuery and a streamArtifactStoreKeys() method to
StoreDataManager. This SHOULD be something that we can optimize, ensuring all keys are available in-memory or easy
to deserialize with minimum overhead. We can do a lot of filtering based on the key alone, then additional
filtering can be applied with the matching stores, which would be retrieved directly by key from the filtered
keys in these calls.