populate the affected-by cache on startup if the affected-by cache is clear
This should work if the affected-by cache is kept in memory, or for the first run when we use a database to
store it. We're lining this up with the by-package-type cache as a single-node fix, where we will still need
a more scalable solution from a proper database / data design approach before we're ready for multi-node usage.