Fix init of affected-by reverse map, and add checks/error message if it's empty (#1488)
* Fix affected-by reversemap initialization in lifecycle startup

* Add logging to note when the affected-by cache is altered, and another error log message if it's empty

* Refactor InfinispanStoreDataManager.init() to startup actions, add some logging

The init() method runs before any migration actions run, which means migrated stores won't be stored in the byPkg cache. The old initializer for the affected-by map has a similar problem, so I removed it in favor of the new startup action. The byPkg init was moved out to a startup action too for this reason.

Finally, I've added a log message to show which groups were found to be affected by a specific set of input keys. This will give us better information about whether the affected-by map is being used or not.

* adding another ftest that initializes store data at boot instead of at migration