Refactoring to fix honeycomb tracing bugs, get threading included in traces (#1508)
* Refactored threaded honeycomb tracing to fix bugs and get proper nesting working

NOTE: This uses a forked version of the beeline-java library currently. I intend to submit a PR and try to get the
new TracingContext merged into the main project. However, until then we need this fork in order to get threaded
tracing.

* TEMPORARILY use a fork of Honeycomb's beeline-java library

I'd like to get the master branch of Indy back up and running without blocking on
https://github.com/honeycombio/beeline-java/pull/28

* don't start the honeycomb stuff if the config has it disabled

* fix NPE when honeycomb is disabled