Adjust the partition key to make the data spread evenly around the cluster
Since there is only 9 partitions if we use the packageType and storeType
as the partition key, that is too less which may bring the hot-node problem.
To avoid it, the basic technique is to add another column to the PRIMARY KEY
to form a compound partition key. Here we add a new column, name_hash_prefix,
holds a prefix of a hash of the store name.For example here:it's the value of
the hash of the first several chars modulo 10. Together with the keys, these two
columns form the compound partition key, and it will be spread across 9*10
partitions.