Refactor schedule expiration to use ispn query api (#1762)
We found that the ScheduleManager.cancel method will trigger the full table scan of the ScheduleCache because of the looping in its stream usage. This caused really bad performance impact on indy-static server. This commit is addressing this problem. It refactor the cache value object from a Map to a pojo which contains that map, and do indexing from ispn level. And then replace the stream looping with ispn query api.