Workaround NPE when downloading a checksum via httprox
When downloading an md5 or a sha1, the metadata object is null for some
reason, but the code does not expect it in a single specific place,
while it does in others. Also setting the contentType to null seems safe,
because it is checked for null further down in the code.

Not sure if it is correct state that metadata is null, but this should
help unblock other features development in PNC.