major overhaul of client metrics, and refactor to use new o11yphant tracing api
The new O11yphant tracing API allows us to plug in Opentelemetry, Honeycomb, or other tracing providers. These changes default to otel (Opentelemetry) for now.

I reworked the client metrics stuff as part of the refactor for the tracing api, but also to avoid storing state about any single request inside ClientMetricsManager. This new approach uses a ClientMetrics object to keep the state confined to the calling method, and is passed in HttpResources when the response needs to be handed back to some caller further up the stack. This should ensure we can make more than one client HTTP call concurrently without polluting or losing metrics.