fix tracing with new otel/honeycomb tracers, wrap jhttpc requests, etc
I've renamed indy-subsys-honeycomb to indy-subsys-trace, since it now works with the generalized tracing API from O11yphant. I added in support (via configuration) for either honeycomb or opentelemetry tracer implementations, and added in the o11yphant-trace-helper-jhttpc to automatically wrap requests with spans, if a tracer is available.

I've updated the client api again, this time to push the span management down into jhttpc, and only do span field additions from the client api itself.

RootSpanFields interface in o11yphant was renamed to SpanFieldsInjector, and can implement before-span or after-span field injection, which is useful for pushing context into thread spans, as well as summarizing metrics into spans after the request is finished (as we've always done with o11yphant).

The new configuration file is trace.conf, not honeycomb.conf.

Finally, I went through some of the points where threads are started inside Indy, and added some span fields to improve context information in those sub-spans where it made sense (like in the MavenMetadataGenerator, for example).

I see that one or more of the unit tests in the scheduler rely on a Cassandra instance. I consider unit tests that depend on external resources to be **broken**, but I haven't disabled them yet. I managed to work around this, but because of this I don't expect a clean build until those are dealt with.