[WIP] Refactor to o11yphant-trace-api, improve trace propagation (#1866)
* Refactor client api metrics implements

* Fix span NPE when header sets

* Refactor client metrics on removing the dependence on o11 honeycomb-core

* Replace dependency of o11 metric-core with metric-common

* Refactor tiny places of client metrics

* Optimize client imports

* Add client metric TRAFFIC_TYPE and adjust way to register some span field

* Refactoring o11yphant to provide generic trace api. WIP

* major overhaul of client metrics, and refactor to use new o11yphant tracing api

The new O11yphant tracing API allows us to plug in Opentelemetry, Honeycomb, or other tracing providers. These changes default to otel (Opentelemetry) for now.

I reworked the client metrics stuff as part of the refactor for the tracing api, but also to avoid storing state about any single request inside ClientMetricsManager. This new approach uses a ClientMetrics object to keep the state confined to the calling method, and is passed in HttpResources when the response needs to be handed back to some caller further up the stack. This should ensure we can make more than one client HTTP call concurrently without polluting or losing metrics.

* fix tracing with new otel/honeycomb tracers, wrap jhttpc requests, etc

I've renamed indy-subsys-honeycomb to indy-subsys-trace, since it now works with the generalized tracing API from O11yphant. I added in support (via configuration) for either honeycomb or opentelemetry tracer implementations, and added in the o11yphant-trace-helper-jhttpc to automatically wrap requests with spans, if a tracer is available.

I've updated the client api again, this time to push the span management down into jhttpc, and only do span field additions from the client api itself.

RootSpanFields interface in o11yphant was renamed to SpanFieldsInjector, and can implement before-span or after-span field injection, which is useful for pushing context into thread spans, as well as summarizing metrics into spans after the request is finished (as we've always done with o11yphant).

The new configuration file is trace.conf, not honeycomb.conf.

Finally, I went through some of the points where threads are started inside Indy, and added some span fields to improve context information in those sub-spans where it made sense (like in the MavenMetadataGenerator, for example).

I see that one or more of the unit tests in the scheduler rely on a Cassandra instance. I consider unit tests that depend on external resources to be **broken**, but I haven't disabled them yet. I managed to work around this, but because of this I don't expect a clean build until those are dealt with.

* Minor bugfixes after cherry-picking the o11yphant tracing refactors over

* Adjust to breakout of helper classes for tracing context propagation

* support for specifying the opentelemetry grpc endpoint URI and export headers

* Add @Alternative to ClientTrafficClassifier to avoid confusion when generating REST api docs

* fix NPE with origin in ctor of HttpResources

* change the default standalone setting back to false.

It appears multiple ftests depend on PathMappedCacheProvider specifically.
'

* use the cassandra plugin to avoid problems when generating the REST API doc

* take out toolchains plugin

This was missed when cherry-picking from master branch

* run swagger export test in standalone mode

* go back and disable all cassandra usage

Also remove Cassandra from the plugin list

* fix tracing with new otel/honeycomb tracers, wrap jhttpc requests, etc

I've renamed indy-subsys-honeycomb to indy-subsys-trace, since it now works with the generalized tracing API from O11yphant. I added in support (via configuration) for either honeycomb or opentelemetry tracer implementations, and added in the o11yphant-trace-helper-jhttpc to automatically wrap requests with spans, if a tracer is available.

I've updated the client api again, this time to push the span management down into jhttpc, and only do span field additions from the client api itself.

RootSpanFields interface in o11yphant was renamed to SpanFieldsInjector, and can implement before-span or after-span field injection, which is useful for pushing context into thread spans, as well as summarizing metrics into spans after the request is finished (as we've always done with o11yphant).

The new configuration file is trace.conf, not honeycomb.conf.

Finally, I went through some of the points where threads are started inside Indy, and added some span fields to improve context information in those sub-spans where it made sense (like in the MavenMetadataGenerator, for example).

I see that one or more of the unit tests in the scheduler rely on a Cassandra instance. I consider unit tests that depend on external resources to be **broken**, but I haven't disabled them yet. I managed to work around this, but because of this I don't expect a clean build until those are dealt with.

* Minor bugfixes after cherry-picking the o11yphant tracing refactors over

* [maven-release-plugin] prepare release indy-parent-2.5.5

* [maven-release-plugin] prepare for next development iteration

* fix tracing with new otel/honeycomb tracers, wrap jhttpc requests, etc

I've renamed indy-subsys-honeycomb to indy-subsys-trace, since it now works with the generalized tracing API from O11yphant. I added in support (via configuration) for either honeycomb or opentelemetry tracer implementations, and added in the o11yphant-trace-helper-jhttpc to automatically wrap requests with spans, if a tracer is available.

I've updated the client api again, this time to push the span management down into jhttpc, and only do span field additions from the client api itself.

RootSpanFields interface in o11yphant was renamed to SpanFieldsInjector, and can implement before-span or after-span field injection, which is useful for pushing context into thread spans, as well as summarizing metrics into spans after the request is finished (as we've always done with o11yphant).

The new configuration file is trace.conf, not honeycomb.conf.

Finally, I went through some of the points where threads are started inside Indy, and added some span fields to improve context information in those sub-spans where it made sense (like in the MavenMetadataGenerator, for example).

I see that one or more of the unit tests in the scheduler rely on a Cassandra instance. I consider unit tests that depend on external resources to be **broken**, but I haven't disabled them yet. I managed to work around this, but because of this I don't expect a clean build until those are dealt with.

* [maven-release-plugin] prepare release indy-parent-2.5.5

* re-bump to next snapshot

* Adding some basic github workflows

* fixing duplicate deps and moving C* fields injector

Co-authored-by: yma <yma@redhat.com>
Co-authored-by: Gang Li <gli@redhat.com>