implement changed-file tracking through transfer decorator
This will improve our ability to target the backup system toward files that have changed.

TODO: we would still need a way to capture changes in the data files, if anything like that happens...such as if promotion rules or autoprox rules changed. Thankfully, these are very uncommon use cases.