File change tracking (#1905)
* Add Indy Client API metrics tracing, incorporate trace_id

* Add return for client API condition

* implement changed-file tracking through transfer decorator

This will improve our ability to target the backup system toward files that have changed.

TODO: we would still need a way to capture changes in the data files, if anything like that happens...such as if promotion rules or autoprox rules changed. Thankfully, these are very uncommon use cases.

* Adding unit test for file-change tracking decorator

* remove some debugging log statements and reorganize methods

Co-authored-by: yma <yma@redhat.com>