Improve metrics / trace measurements to account for transfer times
In the case where tracing is enabled, this change will inject a close-blocking field injector. That allows the main REST call AND the transfer to finish - in any order - before any of the fields are injected. When all the injectors (the REST call and the streaming output) call Span.close(), only THEN will these fields be injected and the span actually get closed.

Metrics are written when the write method closes, and are unaffected by the trace settings