Implement rate/size metrics for transfers and update trace/metric config examples (#1979)
* Implement rate/size metrics for transfers and update trace/metric config examples

* Improve metrics / trace measurements to account for transfer times

In the case where tracing is enabled, this change will inject a close-blocking field injector. That allows the main REST call AND the transfer to finish - in any order - before any of the fields are injected. When all the injectors (the REST call and the streaming output) call Span.close(), only THEN will these fields be injected and the span actually get closed.

Metrics are written when the write method closes, and are unaffected by the trace settings

* fix NPE on construction of TransferStreamingOutput

* Add build feature to allow disabling cluster-oriented tests

This will help when running PR builds in GitHub Actions, etc. where you may not have room to stand up a Cassandra container, or depend on its timing if you do. Using -Pno-cluster-tests will disable cassandra and pathmap tests, in both unit tests and in ftests.

* found two more cases where clustering features are used in ftests.

Adding ability to turn cluster-specific parts off for some PR builds.