Add ability to reset metrics
This adds a new internal feature flag 'test.features.enabled', which is off by default. This flag currently enables a REST endpoint that will reset the metrics on the system. Resetting metrics should allow us to perform different types of performance tests on a running Indy instance, and retrieve performance characteristics that are more or less isolated to that test run, without restarting Indy.