Enable RESET of metrics via REST. Improve latency pause calc for tracing (#1988)
* Move MDCManager + request-based fields into ResourceManagementFilter, other minor tracing fixes

The ThreadContextFilter was outside the TracingFilter, and was adding span fields via RequestContextHelper for request information like client IP address. This caused problems because the ThreadContextFilter would try to add that data to the span after the trace was closed. To fix this, I moved the addition of those span fields into the ResourceManagementFilter and out of ThreadContextFilter.

I cleaned up the close-blocking behavior in the TraceFilter, but also in the TransferStreamingOutput. This is in response to a consolidation within TraceManager for how field injection works, to prevent a forked instance tree on the same basic span between users trying to prevent the span from closing too soon.

I also refactored some field names into constants, and added a read.kb and read.kbps into TransferCountingInputStream.

Finally, I verified that traces and spans are working as expected in Honeycomb.

* Add ability to reset metrics

This adds a new internal feature flag 'test.features.enabled', which is off by default. This flag currently enables a REST endpoint that will reset the metrics on the system. Resetting metrics should allow us to perform different types of performance tests on a running Indy instance, and retrieve performance characteristics that are more or less isolated to that test run, without restarting Indy.

* Accumulate pause value for latency calculation inside the span

Accumulating time spent in transfer (from upstream, or to/from the client) using the span allows us to combine the data from these two classes with latency-pause data from requests we make upstream, in the o11yphant jhttpc helper, and hopefully arrive at more stable internal-to-Indy latency calculations.